Add "Set up a frontend development environment" (12.15)
Turns out the local volume can be shared just fine on Windows too,
though there is a slight hickup which makes Vite unable to fast refresh,
but that can be fixed with enabling polling on the server (see vite config).
Could also supposedly do development entirely in WSL to fix the issue,
but where's the fun in that.

Secondly, I'm only sharing the src file as that's mostly what's
needed for fast refresh. Also, sharing the whole folder is generally
the issue with the whole "missing this-and-that" in node_modules: the install
should install the platform-specific dependencies, but the volume share
causes everything to be copied over -- as I understand it anyway -- and
effectively overwrites whatever the install did, IF node_modules is
present in the shared host folder as well.