Add "Create a production todo-app/docker-compose.yml file" (12.20)
Largely a copy-paste. Main difference is in the frontend also using
nginx, which required different settings for the proxy passing
to the frontend's html port. Additionally, due to the frontend
being built, the environment variable for the backend url cannot
be set after the build (or rather, is not read anymore after the build),
so I had to add that to the Dockerfile instead
of having it in the compose file where setting it does not have an
effect as the build has already completed.