Remove volumes from production build backend and frontend (12.)
The volumes are shared in the dev build because of the fast refresh,
so they're not useful in the production build. Mainly not entirely
sure what would even happen if the referred files weren't there on the host,
though I imagine at worst nothing gets copied (literally replacing
existing files with nothing), and at best nothing happens.